Sietch Jacurutu Level Started (Post-Research)
Spent the day researching the lore of Dune. I have developed a map that contains all of the "necessary" facilities that a community of Fremen (or Iduali) would need to endure Arrakis.

I started laying the foundations of the map. Trying to get scaling right away. I believe much of the map is going to be subtractive, due to numerous underground elements.