Tons of environment refinement, added Spice (Melange) room
Significantly refined the look of the rocky walls, adding tons of cracks, gouges, and major cutaways to make it more "organic" in nature.

The melange room is cut out, finally.

Box collision added so the player cannot leave the playspace.

Lighting touched up and full rebuild.

If I get more time later in the night, I will add more smaller objects (such as tables, beds, and other things to signify machines n such). The corridors and rooms have lofty ceilings, but that is intentional. The smaller objects would really help in interpreting scale for the player.