Starting JourneyAcrossDesert
I had intended to start this project earlier this week, and then some urgent family matters occurred. Going to try pumping this out tonight, albeit not a good idea to do it all in one sitting...

Just finished researching some of the area that I'll need to work with, taking a lot of inspiration from Persian/Iranian ancient dwellings and qanat systems. 

This stage contains some of the building blocks for the ancient city you start in. I will be turning these CSG blockouts into meshes, which will be easily duplicated into a full city.