Lots of refinement, each area laid out
Each area is now in the game, at least to some degree. Areas 1, 2, 3, and 4 are place and laid out functionally. Now to spruce it up and refine the desert where the worm chases the player.

Lots of changes regarding landscaping, the qanat tunnel, the worm, and the landscape parameters had to be addressed. The worm initially would not chase the player, nor could it go underground. All should be operational now.