Wrapping it up, the journey is done
Added collisions so the player is unable to leave the play space! The jetpack and teleport abilities were considered when placing the collision volumes.

Reset some of the worm blueprint stuff, specifically its size and speed. It seems more accurate/deadly at its default settings.

Lastly, the one thing I wanted to change is the projectile's velocity... As of right now, it is possible to skip the jetpack and go straight to the end of the level... I intended on fixing this by changing the velocity of the teleport projectile, but I cannot find it for the life of me. I've been searching for 45 minutes, so I'll save the question for class time.

Also uploading the remainder of the content that was added via starter content and mesh-making. Just to be sure that everything is in sync.